Send email if mail address not verified
This sends a verification email to the provided email address. A user cannot sign in with a new account or updated email address without verifying the provider email address. In addition, if someone provides a fake email address, they cannot receive the link. Consequently, the property `emailVerified` will be `false`. We could delete all accounts that are not verified after a specified period of time.